FIX: Catch error of auth code reuse
Spotify API returns 400 error when an auth code is reused to try to retrieve Tokens again. This is now caught. 
Since this error could also be due to an invalid auth token, 'checkIfAuthCodeFirstUse()' checks whether the auth code was used for a successful presave before. For this reason, auth codes are now also stored in Firestore